requirements.txt
pymupdf
sentence-transformers
psycopg2-binary
asyncpg
pytest
pytest-asyncio
python-dotenv
openai
fastapi
uvicorn
jinja2
pytesseract
pillow
python-multipart
sarvamai
litellm
maturin>=1.5,<2.0
